Reorg history to util
